requirements.txt
pydantic>=2.7
python-dotenv>=1.0
PyYAML>=6.0
tqdm>=4.66
tabulate>=0.9
datasets>=2.20.0
pandas>=2.2.2
numpy>=1.26
scipy>=1.11
matplotlib>=3.8
httpx>=0.27.0
tenacity>=8.3.0
tiktoken>=0.7.0
jsonlines>=4.0.0
openai>=1.33.0
